Cleanup components (#1096)
* Cleanup components
* Renaming spec files
* Mergin non standalone components
* Split tests
