+++ src/components/avatar/Avatar.spec.jsx
