+++ src/components/list/MenuList.jsx
@@ -1,12 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import MenuItem from '../list/MenuItem';
 
 const SIZE = {
   SMALL: 'small',
   NORMAL: 'normal',
   LARGE: 'large'
+};
+
+const MenuItem = ({text, href}) => <li className="sg-menu-list__element">
+  <a className="sg-menu-list__link" href={href}>{text}</a>
+</li>;
+
+MenuItem.propTypes = {
+  href: PropTypes.string.isRequired,
+  text: PropTypes.string.isRequired
 };
 
 const MenuList = ({items, size = SIZE.NORMAL}) => {
@@ -27,4 +35,4 @@ MenuList.propTypes = {
 };
 
 export default MenuList;
-export {SIZE};
+export {SIZE, MenuItem};
