+++ src/components/stickers/Sticker.spec.jsx
